Adding a CounterChain module that can be configured to either run independently or chained together.
This exposes a problem in the infrastructure: config bits for each
counter are stored within the counter module, and config bits regarding
chaining are stored in the counter chain module. With this setup, it is
not possible to 'poke' the 'opcode' port of the counter modules within
the counter chain in Chisel. This means that nested ConfigurableModules
cannot be programmed hierarchically using pokes in simulation.

Configuration could be passed in during module creation. Exploring that
alternative next.