Testing a new way of simulation. Every module will have an 'abstract' version which defines the IO interface and config register, if any. Then, two sub-classes are created. One subclass defines all the actual logic inside the module, the other class is an empty class that pretends to be the actual module by overriding its name. The testbench is defined on the abstract class.
This setup allows us to create the actual module during simulator
creation, but the dummy module during testing. This saves a lot of time
in Chisel compilation. In addition, using the 'null' backend during
testing saves time in creating the output file. We can thus create the
simulator once (at the usual heavy cost) and re-use it for multiple
configurations.

MemoryUnit was a trial, and it seems to work as required. The test is
failing because of a faulty FIFO configuration. This will be fixed
shortly.